Move DependencyFilter out of internal
